Changes to photos, text, and layout to fix look of portfolio.
I added photos for images of skills section in about me, and photos of me. Changed home page layout to have buttons instead of texts, and a design logo. Added home button to each. Changed some spacing and padding. In about me I added, skills and work experience. In designs I didn't like how to the photos were laid out so i tried to make a carousel with the photos but didn't like how it wouldn't restart so I deleted it.